--- handlers/service_migrator.go
@@ -1,15 +1,0 @@
-package handler
-
-import (
-	"getservices/client"
-)
-
-type ServicesHandler struct {
-	Services client.ClientMigrations
-}
-
-func NewServiceHandler(s client.ClientMigrations) ServicesHandler {
-	return ServicesHandler{
-		Services: s,
-	}
-}
