--- server/service_provider.go
@@ -1,15 +1,0 @@
-package server
-
-import (
-	handler "getservices/handlers"
-)
-
-type ServiceProviderServer struct {
-	Provider handler.ServiceProviderHandler
-}
-
-func NewServiceProviderServer(p handler.ServiceProviderHandler) ServiceProviderServer {
-	return ServiceProviderServer{
-		Provider: p,
-	}
-}
