+++ storage/provider_services.go
@@ -1,4 +1,4 @@
-package client
+package storage
 
 import (
 	"getservices/models"
